chore: adding upgrade module
