+++ ignition/deployedGovernorProxy.json
@@ -1,0 +1,5 @@
+{
+  "governorUpgrade": {
+    "governorProxyAddress": "0x29C3c6DCB27580a6dc6fb0c70C797096b73B6d0A"
+  }
+}
