+++ ignition/modules/GovernorUpgradeModule.ts
@@ -1,0 +1,36 @@
+import { buildModule } from '@nomicfoundation/hardhat-ignition/modules';
+
+const governorUpgradeModule = buildModule('GovernorUpgrade', m => {
+  const deployer = m.getAccount(0);
+
+  // Load the governor proxy address
+  const governorProxyAddress = m.getParameter('governorProxyAddress');
+
+  // Use the module to fetch the existing proxy
+  const governorProxyAddr = m.contractAt('RootDao', governorProxyAddress)
+
+  //`RootDaoV2` is the new version
+  const newGovernorImplementation = m.contract('RootDaoV2');
+
+  // Prepare upgrade data
+  const upgradeData = m.encodeFunctionCall(newGovernorImplementation, 'initialize', []);
+
+  // Perform the upgrade
+  const upgradeTx = m.call(governorProxyAddr, 'upgradeToAndCall', [newGovernorImplementation, upgradeData], {
+    from: deployer,
+    id: 'upgrade_governor',
+  });
+
+  console.log(`Governor upgrade tx: ${upgradeTx}`);
+
+  return { governorProxyAddr, newGovernorImplementation };
+});
+
+const governorV2Module = buildModule('GovernorV2', m => {
+  const { governorProxyAddr } = m.useModule(governorUpgradeModule);
+  const gV2 = m.contractAt('RootDaoV2', governorProxyAddr);
+
+  return { gV2 };
+});
+
+export default governorV2Module;
